access-analyzer/access-analyzer-example/access-analyzer-example/requirements.txt
-e .
aws-cdk.core==1.36.1
aws_cdk.aws_s3==1.36.1
aws_cdk.aws_accessanalyzer==1.36.1
aws_cdk.aws_kms==1.36.1
aws_cdk.aws_iam==1.36.1
aws_cdk.aws_lambda==1.36.1
aws_cdk.aws_events==1.36.1
aws_cdk.aws_events_targets==1.36.1
aws_cdk.aws_sqs==1.36.1
aws_cdk.aws_sns==1.36.1
aws_cdk.aws_sns_subscriptions==1.36.1
aws_cdk.aws_stepfunctions==1.36.1
aws_cdk.aws_stepfunctions_tasks==1.36.1
